# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 健身房後端 API — 最終作業規格書
  version: 1.0.0
  description: |
    ## 這本規格書怎麼用

    左側依**里程碑**分組（M1 → M6 就是建議的施工順序）。每支 API 都有
    request / response 範例：先看範例、再看欄位說明，需要登入的 API
    按右上角 **Authorize** 貼上 login 拿到的 token，就能直接 **Try it out** 試打你的後端。

    ## 全域回應格式（每一支都適用）

    | 情況 | HTTP 狀態碼 | Body |
    |---|---|---|
    | 成功 | 2xx | `{ "status": "success", "data": ... }` |
    | 可預期的失敗（驗證錯、找不到、沒權限） | 4xx | `{ "status": "failed", "message": "..." }` |
    | 你的程式壞掉 | 5xx | （不在驗收範圍，但請努力不要發生） |

    - **驗收只看 2xx / 4xx + status 欄位**：文件裡寫的 200/201 是參考值，回 200 或 201 都算對
    - 需要登入的 API 一律收 `Authorization: Bearer <token>`
    - JWT payload 必須含 `{ id, role, exp }`（前端會解開來用，缺了會壞）
    - 所有 id 都是 uuid 字串
    - 日期時間欄位用 UTC ISO 8601 字串（例 `2026-08-20T10:00:00Z`）

    ## ⚠️ 四句固定錯誤訊息文字（全文件最重要的一段）

    以下四句錯誤訊息，前端會拿來判斷要顯示哪個畫面，必須一字不差：

    1. `已經報名過此課程`
    2. `已無可使用堂數`
    3. `已達最大參加人數，無法參加`
    4. `請先登入`

    其他錯誤訊息文字隨你寫（只要 status 是 failed）。
servers:
  - url: http://localhost:8080
    description: 你的後端（本機開發）
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 用 POST /api/users/login（或 POST /api/users/google、/github、/facebook）拿到的 token
paths:
...
  /api/credit-package:
    get:
      tags:
        - "M1 種資料：技能與方案"
      summary: 取得購買方案列表
      description: |-
        回傳全部堂數方案，不需登入。前端的「健身方案」頁直接吃這份資料。
        ⚠️ `price` 的型別：答案版回的是**數字字串**（例如 `"1400.00"`），你回**數字**（例如 `1400`）也可以——前端兩種都處理得了，不構成驗收差異。文件範例照答案版寫字串。
      responses:
        '200':
          description: 成功取得方案列表（還沒有任何方案時回空陣列 `[]`，不是錯誤）
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          description: 方案 id（uuid 字串）
                        name:
                          type: string
                          description: 方案名稱
                        credit_amount:
                          type: integer
                          description: 此方案包含的堂數
                        price:
                          description: 價格。數字或數字字串皆可（答案版回字串 "1400.00"）
                          oneOf:
                            - type: number
                            - type: string
              example:
                status: success
                data:
                  - id: "0d2c1b3a-4e5f-4a6b-8c7d-9e0f1a2b3c4d"
                    name: "7 堂組合包方案"
                    credit_amount: 7
                    price: "1400.00"
                  - id: "1e3d2c4b-5f6a-4b7c-9d8e-0f1a2b3c4d5e"
                    name: "14 堂組合包方案"
                    credit_amount: 14
                    price: "2520.00"
    post:
      tags:
        - "M1 種資料：技能與方案"
      summary: 新增購買方案
      description: |-
        新增一筆堂數方案，不需登入（課程簡化的管理端點；前端頁面沒有呼叫它，但它是把「健身方案」資料種進系統的唯一入口——不先用這支種資料，方案頁會永遠空白、購買流程也無從驗收）。
        方案名稱不可重複。`credit_amount` 與 `price` 都必須是 **0 以上的整數（數字型別）**——price 請送數字，不要送字串，送字串會被擋 400。
        ⚠️ 成功狀態碼是 **200，不是 201**（答案版實際行為）。驗收只看「成功 2xx ＋ status 欄位」，回 201 也算過。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - credit_amount
                - price
              properties:
                name:
                  type: string
                  description: 方案名稱，必填、不可為空字串或全空白、不可與既有方案重複
                credit_amount:
                  type: integer
                  description: 方案包含的堂數，必填、0 以上的整數
                price:
                  type: integer
                  description: 方案價格，必填、0 以上的整數（要送數字型別，不能送字串）
            example:
              name: "7 堂組合包方案"
              credit_amount: 7
              price: 1400
      responses:
        '200':
          description: 新增成功，回傳完整的方案資料（含後端產生的 id 與建立時間）
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        description: 後端產生的 uuid
                      name:
                        type: string
                      credit_amount:
                        type: integer
                      price:
                        type: number
                      createdAt:
                        type: string
                        description: 建立時間（ISO 8601 字串）
              example:
                status: success
                data:
                  id: "0d2c1b3a-4e5f-4a6b-8c7d-9e0f1a2b3c4d"
                  name: "7 堂組合包方案"
                  credit_amount: 7
                  price: 1400
                  createdAt: "2026-06-10T08:00:00.000Z"
        '400':
          description: 觸發條件：任一欄位沒給；name 不是字串或為空；credit_amount 或 price 不是數字、是負數、或帶小數
          content:
            application/json:
              example:
                status: failed
                message: "欄位未填寫正確"
        '409':
          description: 觸發條件：name 與既有方案重複
          content:
            application/json:
              example:
                status: failed
                message: "資料重複"
...
  /api/users/signup:
      post:
        tags:
          - M2 會員系統
        summary: 註冊新會員帳號
        description: |
          建立一個新會員。身分（role）固定是 USER，註冊時無法指定身分。

          行為備註：
          - 密碼規則由後端把關：必須同時包含英文大寫、英文小寫、數字，長度 8～16 字。不符合規則回 400，錯誤訊息可參考下方範例。
          - Email 不可重複，已被註冊過回 409。
          - body 多帶的欄位（例如前端表單順手送來的 confirmPassword）會被忽略；「兩次密碼是否一致」是前端的事，這支 API 不檢查。
          - 成功只回 user 的 id（uuid 字串）與 name，不回 email、絕對不回密碼。
          - ⚠️ 成功狀態碼是 201（不是 200）。驗收尺度：成功＝2xx ＋ body.status 為 "success"、可預期失敗＝4xx ＋ body.status 為 "failed"，文中的數字是實際值供參考。
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                required:
                  - name
                  - email
                  - password
                properties:
                  name:
                    type: string
                    description: 使用者暱稱，不可為空字串或空白
                    example: 小華
                  email:
                    type: string
                    description: 登入用 Email，全站不可重複
                    example: xiaohua@example.com
                  password:
                    type: string
                    description: 必須同時含英文大小寫與數字，8～16 字
                    example: Hexschool12345
              example:
                name: 小華
                email: xiaohua@example.com
                password: Hexschool12345
        responses:
          '201':
            description: 註冊成功（⚠️ 是 201 不是 200）
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    status:
                      type: string
                      example: success
                    data:
                      type: object
                      properties:
                        user:
                          type: object
                          properties:
                            id:
                              type: string
                              description: uuid 字串
                              example: 0fd96b3c-6e2a-4f2b-8d3e-1c2a4b5d6e7f
                            name:
                              type: string
                              example: 小華
                example:
                  status: success
                  data:
                    user:
                      id: 0fd96b3c-6e2a-4f2b-8d3e-1c2a4b5d6e7f
                      name: 小華
          '400':
            description: |
              欄位驗證失敗。下方是建議回應範例；驗收重點是 4xx + status: failed。
              - name、email、password 任一缺漏或為空字串：「欄位未填寫正確」
              - 密碼不符合規則（缺大寫／小寫／數字、或長度不在 8～16）：「密碼不符合規則，需要包含英文數字大小寫，最短8個字，最長16個字」
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    status:
                      type: string
                      example: failed
                    message:
                      type: string
                examples:
                  欄位缺漏:
                    summary: 任一必填欄位缺漏或為空
                    value:
                      status: failed
                      message: 欄位未填寫正確
                  密碼不符規則:
                    summary: 密碼不符合大小寫英數 8~16 字規則
                    value:
                      status: failed
                      message: 密碼不符合規則，需要包含英文數字大小寫，最短8個字，最長16個字
          '409':
            description: Email 已經被別人註冊過
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    status:
                      type: string
                      example: failed
                    message:
                      type: string
                example:
                  status: failed
                  message: Email 已被使用
  /api/users/login:
    post:
      tags:
        - M2 會員系統
      summary: 會員登入，取得 JWT token
      description: |
        用 Email 與密碼登入，成功會拿到一組 token，之後所有需要登入的 API 都把它放在
        Authorization header（格式：Bearer <token>）一起送出。

        行為備註：
        - ⚠️ JWT payload 格式：解開 token 後必須包含 { id, role, exp } 三個欄位
          （id 是使用者 uuid、role 是 "USER" 或 "COACH"、exp 是過期時間）。前端會自己解析 token
          讀這三個欄位來決定顯示哪些選單，少任何一個前端就會壞掉。
        - ⚠️ 「帳號不存在」與「密碼錯誤」共用同一句 400 訊息「使用者不存在或密碼輸入錯誤」——
          刻意不告訴別人這個 Email 有沒有註冊過，兩種情況都要回一模一樣的訊息。
        - 登入時也會先檢查欄位與密碼規則：密碼格式根本不符合規則時直接回 400（連帳號都不會去查）。
        - 成功回 data.token 與 data.user.name（只有 name，沒有其他欄位）。
        - ⚠️ 成功狀態碼是 201（不是 200）。驗收尺度：只看 2xx／4xx ＋ status 欄位，數字供參考。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  description: 註冊時使用的 Email
                  example: xiaohua@example.com
                password:
                  type: string
                  description: 登入密碼
                  example: Hexschool12345
            example:
              email: xiaohua@example.com
              password: Hexschool12345
      responses:
        '201':
          description: 登入成功（⚠️ 是 201 不是 200）
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        description: JWT token，payload 必含 { id, role, exp }
                        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjBmZDk2YjNjIiwicm9sZSI6IlVTRVIiLCJleHAiOjE3NTAwMDAwMDB9.signature
                      user:
                        type: object
                        properties:
                          name:
                            type: string
                            example: 小華
              example:
                status: success
                data:
                  token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjBmZDk2YjNjIiwicm9sZSI6IlVTRVIiLCJleHAiOjE3NTAwMDAwMDB9.signature
                  user:
                    name: 小華
        '400':
          description: |
            登入失敗。下方是建議回應範例；驗收重點是 4xx + status: failed。
            - email 或 password 缺漏／空字串：「欄位未填寫正確」
            - 密碼格式不符合規則（此時不會去查帳號）：「密碼不符合規則，需要包含英文數字大小寫，最短8個字，最長16個字」
            - 帳號不存在「或」密碼比對錯誤（⚠️ 兩種情況共用同一句）：「使用者不存在或密碼輸入錯誤」
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: failed
                  message:
                    type: string
              examples:
                欄位缺漏:
                  summary: email 或 password 缺漏／空字串
                  value:
                    status: failed
                    message: 欄位未填寫正確
                密碼不符規則:
                  summary: 密碼格式不符合規則
                  value:
                    status: failed
                    message: 密碼不符合規則，需要包含英文數字大小寫，最短8個字，最長16個字
                帳密錯誤:
                  summary: 帳號不存在或密碼錯誤（共用訊息）
                  value:
                    status: failed
                    message: 使用者不存在或密碼輸入錯誤
  /api/users/google:
    post:
      tags:
        - M2 會員系統
      summary: Google 登入，取得 JWT token（僅 Spring Boot 版）
      description: |
        用 Google 帳號登入。前端透過 Google Identity Services 取得 ID token（credential），
        送到這支 API 由後端驗證，成功後回傳的內容與 POST /api/users/login 完全相同，
        之後的用法也一樣（放在 Authorization header：Bearer <token>）。

        行為備註：
        - ⚠️ 這支 API 只有 Spring Boot 版後端（livefit/）有實作，Node 版（backend/）沒有，不列入作業驗收。
        - 後端會用 Google 的公鑰驗證 credential 的簽章，並檢查發行者（iss）、
          對象（aud 必須是本專案的 Google Client ID）、有效期限（exp），且 email 必須是 Google 驗證過的（email_verified）。
        - 帳號對應規則（依序判斷）：
          1. 這個 Google 帳號以前登入過 → 直接登入原帳號。
          2. 沒登入過，但 Google 的 Email 已經有人用密碼註冊 → 自動綁定到那個既有帳號並登入。
             綁定不會動到原本的密碼，之後密碼登入與 Google 登入都能用，也仍然可以修改密碼。
          3. 都沒有 → 建立新會員（role 固定是 USER、沒有密碼），暱稱取 Google 帳號的名稱（最多 50 字）。
        - ⚠️ 用 Google 建立的帳號沒有密碼：不能用 POST /api/users/login 登入，也不能用 PUT /api/users/password 修改密碼。
        - JWT payload 與一般登入相同，必含 { id, role, exp }。
        - 成功回 data.token 與 data.user.name（只有 name，沒有其他欄位）。
        - 成功狀態碼是 201，與一般登入一致。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - credential
              properties:
                credential:
                  type: string
                  description: Google Identity Services 回傳的 ID token（JWT 字串）
                  example: eyJhbGciOiJSUzI1NiIsImtpZCI6IjEyMzQ1Njc4OTAiLCJ0eXAiOiJKV1QifQ.eyJpc3MiOiJodHRwczovL2FjY291bnRzLmdvb2dsZS5jb20ifQ.signature
            example:
              credential: eyJhbGciOiJSUzI1NiIsImtpZCI6IjEyMzQ1Njc4OTAiLCJ0eXAiOiJKV1QifQ.eyJpc3MiOiJodHRwczovL2FjY291bnRzLmdvb2dsZS5jb20ifQ.signature
      responses:
        '201':
          description: 登入成功，回應格式與 POST /api/users/login 相同
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        description: 本站的 JWT token（不是 Google 的 token），payload 必含 { id, role, exp }
                        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjBmZDk2YjNjIiwicm9sZSI6IlVTRVIiLCJleHAiOjE3NTAwMDAwMDB9.signature
                      user:
                        type: object
                        properties:
                          name:
                            type: string
                            example: 小華
              example:
                status: success
                data:
                  token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjBmZDk2YjNjIiwicm9sZSI6IlVTRVIiLCJleHAiOjE3NTAwMDAwMDB9.signature
                  user:
                    name: 小華
        '400':
          description: |
            登入失敗。
            - credential 缺漏／空字串：「欄位未填寫正確」
            - credential 不是合法的 Google ID token（簽章錯誤、已過期、不是發給本專案的、Email 未經 Google 驗證）：「Google 登入驗證失敗」
            - 後端沒有設定 GOOGLE_CLIENT_ID：「尚未設定 Google 登入」
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: failed
                  message:
                    type: string
              examples:
                欄位缺漏:
                  summary: credential 缺漏／空字串
                  value:
                    status: failed
                    message: 欄位未填寫正確
                驗證失敗:
                  summary: credential 無法通過驗證
                  value:
                    status: failed
                    message: Google 登入驗證失敗
                尚未設定:
                  summary: 後端未設定 Google Client ID
                  value:
                    status: failed
                    message: 尚未設定 Google 登入
        '409':
          description: 這個 Email 的帳號已經綁定了另一個 Google 帳號
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: failed
                  message:
                    type: string
              example:
                status: failed
                message: Email 已被使用
  /api/users/github:
    post:
      tags:
        - M2 會員系統
      summary: GitHub 登入，取得 JWT token（僅 Spring Boot 版）
      description: |
        用 GitHub 帳號登入。前端把使用者導向 GitHub 授權頁，授權後 GitHub 帶著一次性的授權碼（code）
        跳回前端的 /oauth/callback/github，前端再把 code 送到這支 API。
        成功後回傳的內容與 POST /api/users/login 完全相同，之後的用法也一樣。

        行為備註：
        - ⚠️ 這支 API 只有 Spring Boot 版後端（livefit/）有實作，Node 版（backend/）沒有，不列入作業驗收。
        - 後端會拿 code 加上 Client secret 向 GitHub 換 access token，再用它取得使用者的 id、名稱與 email。
          access token 用完即丟，不會儲存，也不會回傳給前端。
        - redirect_uri 必須是後端 OAUTH_REDIRECT_URIS 白名單內的網址，而且要與前端導向授權頁時用的完全相同。
        - 只採用 GitHub 上「主要且驗證過」的 email；帳號沒有這樣的 email 就無法登入。
        - 帳號對應規則與 Google 登入相同（依序判斷）：
          1. 這個 GitHub 帳號以前登入過 → 直接登入原帳號。
          2. 沒登入過，但該 email 已經有帳號（密碼註冊或用 Google 建立的都算）→ 自動綁定到那個既有帳號並登入，
             原本的密碼與其他登入方式都不受影響。
          3. 都沒有 → 建立新會員（role 固定是 USER、沒有密碼），暱稱取 GitHub 的顯示名稱，沒填的話用帳號名稱（最多 50 字）。
        - ⚠️ 用 GitHub 建立的帳號沒有密碼：不能用 POST /api/users/login 登入，也不能用 PUT /api/users/password 修改密碼。
        - ⚠️ code 只能使用一次，而且約 10 分鐘後過期；重送同一個 code 會得到「GitHub 登入驗證失敗」。
        - JWT payload 與一般登入相同，必含 { id, role, exp }。
        - 成功狀態碼是 201，與一般登入一致。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
                - redirect_uri
              properties:
                code:
                  type: string
                  description: GitHub 授權後帶回前端的一次性授權碼
                  example: 4f3c2b1a0e9d8c7b6a5f
                redirect_uri:
                  type: string
                  description: 前端導向 GitHub 授權頁時使用的 redirect_uri
                  example: http://localhost:5173/oauth/callback/github
            example:
              code: 4f3c2b1a0e9d8c7b6a5f
              redirect_uri: http://localhost:5173/oauth/callback/github
      responses:
        '201':
          description: 登入成功，回應格式與 POST /api/users/login 相同
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        description: 本站的 JWT token（不是 GitHub 的 access token），payload 必含 { id, role, exp }
                        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjBmZDk2YjNjIiwicm9sZSI6IlVTRVIiLCJleHAiOjE3NTAwMDAwMDB9.signature
                      user:
                        type: object
                        properties:
                          name:
                            type: string
                            example: 小華
              example:
                status: success
                data:
                  token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjBmZDk2YjNjIiwicm9sZSI6IlVTRVIiLCJleHAiOjE3NTAwMDAwMDB9.signature
                  user:
                    name: 小華
        '400':
          description: |
            登入失敗。
            - code 缺漏／空字串，或 redirect_uri 缺漏／不在白名單內：「欄位未填寫正確」
            - 後端沒有設定 GITHUB_CLIENT_ID 或 GITHUB_CLIENT_SECRET：「尚未設定 GitHub 登入」
            - code 無效、已使用過或已過期，或連不上 GitHub：「GitHub 登入驗證失敗」
            - GitHub 帳號沒有主要且驗證過的 email：「此 GitHub 帳號沒有已驗證的 Email，無法登入」
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: failed
                  message:
                    type: string
              examples:
                欄位錯誤:
                  summary: code 或 redirect_uri 缺漏，或 redirect_uri 不在白名單
                  value:
                    status: failed
                    message: 欄位未填寫正確
                尚未設定:
                  summary: 後端未設定 GitHub Client ID / secret
                  value:
                    status: failed
                    message: 尚未設定 GitHub 登入
                驗證失敗:
                  summary: code 無法換到使用者資料
                  value:
                    status: failed
                    message: GitHub 登入驗證失敗
                沒有已驗證Email:
                  summary: GitHub 帳號沒有主要且驗證過的 email
                  value:
                    status: failed
                    message: 此 GitHub 帳號沒有已驗證的 Email，無法登入
        '409':
          description: 這個 Email 的帳號已經綁定了另一個 GitHub 帳號
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: failed
                  message:
                    type: string
              example:
                status: failed
                message: Email 已被使用
  /api/users/facebook:
    post:
      tags:
        - M2 會員系統
      summary: Facebook 登入，取得 JWT token（僅 Spring Boot 版）
      description: |
        用 Facebook 帳號登入。流程與 POST /api/users/github 相同：前端把使用者導向 Facebook 授權頁，
        授權後 Facebook 帶著一次性的授權碼（code）跳回前端的 /oauth/callback/facebook，前端再把 code 送到這支 API。
        成功後回傳的內容與 POST /api/users/login 完全相同。

        行為備註：
        - ⚠️ 這支 API 只有 Spring Boot 版後端（livefit/）有實作，Node 版（backend/）沒有，不列入作業驗收。
        - 後端會拿 code 加上應用程式密鑰向 Facebook 換 access token，再用它取得使用者的 id、名稱與 email。
          access token 用完即丟，不會儲存，也不會回傳給前端。
        - redirect_uri 必須是後端 OAUTH_REDIRECT_URIS 白名單內的網址，而且要與前端導向授權頁時用的完全相同。
        - ⚠️ 與 Google、GitHub 不同：Facebook 不會告訴我們 email 有沒有驗證過，所以**不會自動綁定既有帳號**。
          帳號對應規則（依序判斷）：
          1. 這個 Facebook 帳號以前登入過 → 直接登入原帳號。
          2. 沒登入過，而該 email 已經有帳號（不論是密碼註冊、Google 或 GitHub 建立的）→ 409，請使用者改用原本的方式登入。
          3. 都沒有 → 建立新會員（role 固定是 USER、沒有密碼），暱稱取 Facebook 的名稱（最多 50 字）。
        - ⚠️ Facebook 帳號沒有 email（用手機號碼註冊），或使用者在授權頁取消提供 email 時無法登入。
        - ⚠️ 用 Facebook 建立的帳號沒有密碼：不能用 POST /api/users/login 登入，也不能用 PUT /api/users/password 修改密碼。
        - ⚠️ code 只能使用一次；重送同一個 code 會得到「Facebook 登入驗證失敗」。
        - JWT payload 與一般登入相同，必含 { id, role, exp }。
        - 成功狀態碼是 201，與一般登入一致。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
                - redirect_uri
              properties:
                code:
                  type: string
                  description: Facebook 授權後帶回前端的一次性授權碼
                  example: AQBx7kZ3vYq2mN8pL5tR
                redirect_uri:
                  type: string
                  description: 前端導向 Facebook 授權頁時使用的 redirect_uri
                  example: http://localhost:5173/oauth/callback/facebook
            example:
              code: AQBx7kZ3vYq2mN8pL5tR
              redirect_uri: http://localhost:5173/oauth/callback/facebook
      responses:
        '201':
          description: 登入成功，回應格式與 POST /api/users/login 相同
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        description: 本站的 JWT token（不是 Facebook 的 access token），payload 必含 { id, role, exp }
                        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjBmZDk2YjNjIiwicm9sZSI6IlVTRVIiLCJleHAiOjE3NTAwMDAwMDB9.signature
                      user:
                        type: object
                        properties:
                          name:
                            type: string
                            example: 小華
              example:
                status: success
                data:
                  token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjBmZDk2YjNjIiwicm9sZSI6IlVTRVIiLCJleHAiOjE3NTAwMDAwMDB9.signature
                  user:
                    name: 小華
        '400':
          description: |
            登入失敗。
            - code 缺漏／空字串，或 redirect_uri 缺漏／不在白名單內：「欄位未填寫正確」
            - 後端沒有設定 FACEBOOK_APP_ID 或 FACEBOOK_APP_SECRET：「尚未設定 Facebook 登入」
            - code 無效、已使用過或已過期，或連不上 Facebook：「Facebook 登入驗證失敗」
            - Facebook 帳號沒有 email，或使用者沒有同意提供 email：「此 Facebook 帳號沒有提供 Email，無法登入」
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: failed
                  message:
                    type: string
              examples:
                欄位錯誤:
                  summary: code 或 redirect_uri 缺漏，或 redirect_uri 不在白名單
                  value:
                    status: failed
                    message: 欄位未填寫正確
                尚未設定:
                  summary: 後端未設定 Facebook 應用程式編號 / 密鑰
                  value:
                    status: failed
                    message: 尚未設定 Facebook 登入
                驗證失敗:
                  summary: code 無法換到使用者資料
                  value:
                    status: failed
                    message: Facebook 登入驗證失敗
                沒有Email:
                  summary: Facebook 沒有提供 email
                  value:
                    status: failed
                    message: 此 Facebook 帳號沒有提供 Email，無法登入
        '409':
          description: 這個 Email 已經有帳號。Facebook 的 email 沒有驗證旗標，不會自動綁定
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: failed
                  message:
                    type: string
              example:
                status: failed
                message: 此 Email 已註冊，請改用原本的方式登入
...